# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JAC Hub API - Panel de gestión para estudio indie de videojuegos
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
  recordarSesion?: boolean;
}

export interface UsuarioActual {
  id: number;
  nombre: string;
  email: string;
  rol: string;
  /** @nullable */
  avatar: string | null;
  enLinea?: boolean;
}

export interface AuthResponse {
  usuario: UsuarioActual;
  token: string;
}

export interface ProgresoProyecto {
  nombre: string;
  progreso: number;
}

export interface BugsPorPrioridad {
  prioridad: string;
  cantidad: number;
}

export interface ActividadDia {
  dia: string;
  tareas: number;
  bugs: number;
}

export interface DashboardStats {
  proyectosActivos: number;
  miembrosConectados: number;
  bugsAbiertos: number;
  tareasCompletadasSemana: number;
  totalTareas: number;
  totalBuilds: number;
  progresoPorProyecto?: ProgresoProyecto[];
  bugsPorPrioridad?: BugsPorPrioridad[];
  actividadSemanal?: ActividadDia[];
}

export interface ActividadItem {
  id: number;
  mensaje: string;
  tipo: string;
  fecha: string;
  usuario: string;
  /** @nullable */
  avatar?: string | null;
}

export interface Proyecto {
  id: number;
  nombre: string;
  estado: string;
  progreso: number;
  descripcion: string;
  fechaCreacion: string;
  fechaActualizacion: string;
  miembros?: string[];
  tecnologias?: string[];
}

export interface ProyectoInput {
  nombre: string;
  descripcion: string;
  estado?: string;
  tecnologias?: string[];
}

export interface ProyectoUpdate {
  nombre?: string;
  descripcion?: string;
  estado?: string;
  progreso?: number;
  tecnologias?: string[];
}

export interface Tarea {
  id: number;
  titulo: string;
  /** @nullable */
  descripcion?: string | null;
  columna: string;
  prioridad: string;
  /** @nullable */
  proyectoId?: number | null;
  /** @nullable */
  asignadoA?: string | null;
  etiquetas?: string[];
  fechaCreacion: string;
  /** @nullable */
  fechaLimite?: string | null;
}

export interface TareaInput {
  titulo: string;
  descripcion?: string;
  columna: string;
  prioridad?: string;
  proyectoId?: number;
  asignadoA?: string;
  etiquetas?: string[];
  fechaLimite?: string;
}

export interface TareaUpdate {
  titulo?: string;
  descripcion?: string;
  columna?: string;
  prioridad?: string;
  asignadoA?: string;
  etiquetas?: string[];
  fechaLimite?: string;
}

export interface Bug {
  id: number;
  titulo: string;
  /** @nullable */
  descripcion?: string | null;
  prioridad: string;
  estado: string;
  /** @nullable */
  proyectoId?: number | null;
  /** @nullable */
  asignadoA?: string | null;
  reportadoPor?: string;
  fechaReporte: string;
  /** @nullable */
  fechaResolucion?: string | null;
}

export interface BugInput {
  titulo: string;
  descripcion?: string;
  prioridad: string;
  proyectoId?: number;
  asignadoA?: string;
}

export interface BugUpdate {
  titulo?: string;
  descripcion?: string;
  prioridad?: string;
  estado?: string;
  asignadoA?: string;
  fechaResolucion?: string;
}

export interface Miembro {
  id: number;
  nombre: string;
  rol: string;
  email: string;
  /** @nullable */
  avatar?: string | null;
  enLinea: boolean;
  habilidades?: string[];
  fechaUnion: string;
  proyectosActivos?: number;
}

export interface MiembroInput {
  nombre: string;
  rol: string;
  email: string;
  habilidades?: string[];
}

export interface MiembroUpdate {
  nombre?: string;
  rol?: string;
  email?: string;
  enLinea?: boolean;
  habilidades?: string[];
}

export interface Build {
  id: number;
  version: string;
  estado: string;
  /** @nullable */
  descripcion?: string | null;
  /** @nullable */
  changelog?: string | null;
  /** @nullable */
  proyectoId?: number | null;
  subidoPor?: string;
  fechaSubida: string;
  tamano: string;
  plataforma?: string;
  descargas?: number;
}

export interface BuildInput {
  version: string;
  descripcion?: string;
  changelog?: string;
  proyectoId?: number;
  plataforma: string;
  tamano?: string;
}

export interface BuildUpdate {
  estado?: string;
  changelog?: string;
  descripcion?: string;
}

export interface Notificacion {
  id: number;
  mensaje: string;
  tipo: string;
  leida: boolean;
  fecha: string;
  /** @nullable */
  enlace?: string | null;
  /** @nullable */
  usuario?: string | null;
}

export type ListTareasParams = {
  proyectoId?: number;
  columna?: string;
};

export type ListBugsParams = {
  proyectoId?: number;
  prioridad?: string;
  estado?: string;
};

export type ListBuildsParams = {
  proyectoId?: number;
};