# lib/api-zod/src/generated/types/usuarioActual.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JAC Hub API - Panel de gestión para estudio indie de videojuegos
 * OpenAPI spec version: 0.1.0
 */

export interface UsuarioActual {
  id: number;
  nombre: string;
  email: string;
  rol: string;
  /** @nullable */
  avatar: string | null;
  enLinea?: boolean;
}